Update vec2 for ivec2
There are cases where floating point coordinates for vectors aren't required (e.g. Mouse) - in those instances we can replace vec2 for ivec2 (integers)

+++ src/game/mouse.cpp
@@ -2,6 +2,7 @@
 #include <cmath>
 #include <SDL2/SDL_image.h>
 #include <glm/glm.hpp>
+#include <glm/ext/vector_int1.hpp>
 #include "spdlog/spdlog.h"
 
 #include "mouse.h"
@@ -22,7 +23,7 @@ Mouse::~Mouse() {
 }
 
 // Query the pixel colour for a pixel on the mousemap at position x, y
-SDL_Color Mouse::mousemap_pixel_colour(const glm::vec2& pixel_offset) const {
+SDL_Color Mouse::mousemap_pixel_colour(const glm::ivec2& pixel_offset) const {
     const unsigned int bpp {mousemap->format->BytesPerPixel};
     
     // TODO: understand wtf happens here
@@ -49,35 +50,35 @@ SDL_Color Mouse::mousemap_pixel_colour(const glm::vec2& pixel_offset) const {
 }
 
 // Convert the pixel colour into a vector to be added to a 'coarse' grid location
-glm::vec2 Mouse::pixel_colour_vector(const SDL_Colour& colour) const {
+glm::ivec2 Mouse::pixel_colour_vector(const SDL_Colour& colour) const {
 
     if (colour.r == 255 && colour.g == 255 && colour.b == 255) {
-        return glm::vec2{0, 0};
+        return glm::ivec2{0, 0};
     }
 
     if (colour.r == 255) {
-        return glm::vec2{-1, 0};
+        return glm::ivec2{-1, 0};
     }
 
     if (colour.g == 255) {
-        return glm::vec2{0, -1};
+        return glm::ivec2{0, -1};
     }
 
     if (colour.b == 255) {
-        return glm::vec2{1, 0};
+        return glm::ivec2{1, 0};
     }
 
     if (colour.r == 0 && colour.g == 0 && colour.b == 0) {
-        return glm::vec2{0, 1};
+        return glm::ivec2{0, 1};
     }
 
-    return glm::vec2{0, 0};
+    return glm::ivec2{0, 0};
 }
 
 // Calculate the 'coarse' grid position tile walk map
-glm::vec2 Mouse::tile_walk(const glm::vec2& tile_offset) const {
-    glm::vec2 vertical{0, 0};
-    glm::vec2 horizontal{0, 0};
+glm::ivec2 Mouse::tile_walk(const glm::ivec2& tile_offset) const {
+    glm::ivec2 vertical{0, 0};
+    glm::ivec2 horizontal{0, 0};
 
     if (tile_offset.x != 0) {
         horizontal = horizontal_vector * tile_offset.x;
@@ -92,7 +93,7 @@ glm::vec2 Mouse::tile_walk(const glm::vec2& tile_offset) const {
 }
 
 // Public function converting x, y screen coordinates into tilemap coordinates
-glm::vec2 Mouse::pixel_to_grid() const {
+glm::ivec2 Mouse::pixel_to_grid() const {
     // Coarse coordinates
     int screen_offset_x {static_cast<int>(position.x) - constants::TILEMAP_X_START};
     int screen_offset_y {static_cast<int>(position.y) - constants::TILEMAP_Y_START};
@@ -134,10 +135,10 @@ glm::vec2 Mouse::pixel_to_grid() const {
         remainder_y = screen_offset_y % constants::TILE_HEIGHT;
     }
 
-    glm::vec2 coarse {tile_walk({tile_offset_x, tile_offset_y})};
+    glm::ivec2 coarse {tile_walk({tile_offset_x, tile_offset_y})};
 
     SDL_Color pixel_colour = mousemap_pixel_colour(
-        glm::vec2{remainder_x, remainder_y}
+        glm::ivec2{remainder_x, remainder_y}
     );
 
     return coarse + pixel_colour_vector(pixel_colour);
@@ -145,11 +146,13 @@ glm::vec2 Mouse::pixel_to_grid() const {
 
 void Mouse::update() {
     previous_position = position;
+
     int x, y;
     SDL_GetMouseState(&x, &y);
     position.x = x;
     position.y = y;
-    glm::vec2 grid_position {pixel_to_grid()};
+
+    glm::ivec2 grid_position {pixel_to_grid()};
 
     if (position != previous_position) {
         spdlog::info(
